Crate bevy_has_load_progress

Helper plugin and derive macro for tracking the load progress of nested structs containing asset handles.

This crate exposes a trait HasLoadProgress that may be derived on structs that contain Bevy asset [Handle]s. The idea is that you may have a struct with asset handles contained somewhere inside, maybe deeply nested or stored in vectors, etc., and you need to be able to get the load progress of all of the handles inside that struct.

Macros

impl_default_load_progress 🔒

Structs

LoadProgress

A progress indicator holding how many items must be loaded and how many items have been loaded

LoadingResources

System param containing Bevy resources that may be used to determine load progress

Traits

HasLoadProgress

Trait implemented on items that can report their load progress from the LoadingResources.

Derive Macros

HasLoadProgress

Derive macro for HasLoadProgress