Derive Macro bevy_has_load_progress::HasLoadProgress

#[derive(HasLoadProgress)]
{
    // Attributes available to this derive:
    #[has_load_progress]
}

Derive macro for HasLoadProgress

May be used to implement HasLoadProgress on structs where all fields implement HasLoadProgress.

Fields not implementing HasLoadProgress may be skipped with #[has_load_progress(none)] added to the field.

#[has_load_progress(none)] may also be added to the struct itself to use the default implementation of HasLoadProgress which returns no progress and nothing to load.

#[has_load_progress(none)] could also be added to enums to derive the default implementation with no load progress, but cannot be added to enum variants.